Refactor
Refactor concurrency.
Add utils.
Fix gorm's self many2many problem with user model.

+++ model/file.go
@@ -6,8 +6,8 @@ import (
 
 // File is a file model that contains meta.
 type File struct {
-	Id        int64     `json:"id"`
-	UserId    int64     `json:"userId"`
+	Id        uint      `json:"id"`
+	UserId    uint      `json:"userId"`
 	Name      string    `json:"name",sql:"size:255"`
 	Size      int       `json:"size",sql:"size:255"`
 	CreatedAt time.Time `json:"createdAt"`
