+++ service/commentService/form.go
@@ -2,14 +2,14 @@ package commentService
 
 // CreateCommentForm is used when creating a comment.
 type CreateCommentForm struct {
-	UserId   int64  `form:"userId" binding:"required"`
-	ParentId int64  `form:"parentId" binding:"required"`
+	UserId   uint   `form:"userId" binding:"required"`
+	ParentId uint   `form:"parentId" binding:"required"`
 	Content  string `form:"content" binding:"required"`
 }
 
 // CommentForm is used when updating a comment.
 type CommentForm struct {
-	CommentId int64  `form:"commentId" binding:"required"`
-	ParentId  int64  `form:"parentId" binding:"required"`
+	CommentId uint   `form:"commentId" binding:"required"`
+	ParentId  uint   `form:"parentId" binding:"required"`
 	Content   string `form:"content"`
 }
