+++ service/likingService/form.go
@@ -2,6 +2,6 @@ package likingService
 
 // CreateLikingForm is used when creating a liking.
 type CreateLikingForm struct {
-	UserId   int64 `form:"userId" binding:"required"`
-	ParentId int64 `form:"parentId" binding:"required"`
+	UserId   uint `form:"userId" binding:"required"`
+	ParentId uint `form:"parentId" binding:"required"`
 }
